[APM] optimize trace errors terms agg with execution hint (#52726)
* [APM] optimize trace errors terms agg with execution hint

* Use execution_hint: map for transaction group fetcher

* Split up local filter aggregations
