+++ x-pack/legacy/plugins/apm/server/lib/errors/get_trace_errors_per_transaction.ts
@@ -45,12 +45,14 @@ export async function getTraceErrorsPerTransaction(
       aggs: {
         transactions: {
           terms: {
-            field: TRANSACTION_ID
+            field: TRANSACTION_ID,
+            // high cardinality
+            execution_hint: 'map'
           }
         }
       }
     }
-  };
+  } as const;
 
   const resp = await client.search(params);
