+++ x-pack/legacy/plugins/apm/server/lib/ui_filters/local_ui_filters/index.ts
@@ -4,11 +4,10 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import { cloneDeep, sortByOrder } from 'lodash';
-import { mergeProjection } from '../../../../common/projections/util/merge_projection';
 import { UIFilters } from '../../../../typings/ui-filters';
 import { Projection } from '../../../../common/projections/typings';
 import { PromiseReturnType } from '../../../../typings/common';
-import { getFilterAggregations } from './get_filter_aggregations';
+import { getLocalFilterQuery } from './get_local_filter_query';
 import { Setup } from '../../helpers/setup_request';
 import { localUIFilters, LocalUIFilterName } from './config';
 
@@ -33,47 +32,35 @@ export async function getLocalUIFilters({
 
   delete projectionWithoutAggs.body.aggs;
 
-  const filterAggregations = getFilterAggregations({
-    indexPattern: dynamicIndexPattern,
-    uiFilters,
-    projection,
-    localFilterNames
-  });
+  return Promise.all(
+    localFilterNames.map(async name => {
+      const query = getLocalFilterQuery({
+        indexPattern: dynamicIndexPattern,
+        uiFilters,
+        projection,
+        localUIFilterName: name
+      });
 
-  const params = mergeProjection(projectionWithoutAggs, {
-    body: {
-      size: 0,
-      // help TS infer aggregations by making all aggregations required
-      aggs: filterAggregations as Required<typeof filterAggregations>
-    }
-  });
+      const response = await client.search(query);
 
-  const response = await client.search(params);
-  const { aggregations } = response;
+      const filter = localUIFilters[name];
 
-  if (!aggregations) {
-    return [];
-  }
-
-  return localFilterNames.map(key => {
-    const aggregationsForFilter = aggregations[key];
-    const filter = localUIFilters[key];
-
-    return {
-      ...filter,
-      options: sortByOrder(
-        aggregationsForFilter.by_terms.buckets.map(bucket => {
-          return {
-            name: bucket.key as string,
-            count:
-              'bucket_count' in bucket
-                ? bucket.bucket_count.value
-                : bucket.doc_count
-          };
-        }),
-        'count',
-        'desc'
-      )
-    };
-  });
+      return {
+        ...filter,
+        options: sortByOrder(
+          response.aggregations?.by_terms.buckets.map(bucket => {
+            return {
+              name: bucket.key as string,
+              count:
+                'bucket_count' in bucket
+                  ? bucket.bucket_count.value
+                  : bucket.doc_count
+            };
+          }) || [],
+          'count',
+          'desc'
+        )
+      };
+    })
+  );
 }
